paig-server/backend/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
greenlet==3.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
uvicorn==0.30.1
pytest==8.2.2
Faker==25.8.0
pytest-asyncio==0.23.7
pytz==2024.1
coverage==7.5.4
opensearch-py==2.6.0
jproperties==2.1.1
urllib3==2.0.6
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1.tar.gz
cryptography==41.0.4
boto3==1.34.20
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-mock==3.12.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
paig_common
alt-profanity-check==1.5.0
numpy==1.26.4
scipy==1.14.1
starlette-request-id==0.7.0
httpx==0.27.0
fasteners
APScheduler==3.10.4
pydantic[email]
posthog
paig-authorizer-core
click==8.1.7
psutil